Creates new ESGF and CDAT apps for all ESGF/CDAT related endpoints
