+++ apps/util/utilities.py
@@ -1,0 +1,19 @@
+def get_client_ip(request):
+    """see: http://stackoverflow.com/a/4581997"""
+    x_forwarded_for = request.META.get('HTTP_X_FORWARDED_FOR')
+    if x_forwarded_for:
+        ip = x_forwarded_for
+    else:
+        ip = request.META.get('REMOTE_ADDR')
+    return ip
+
+
+def print_debug(e):
+    import traceback
+    print '1', e.__doc__
+    print '2', sys.exc_info()
+    print '3', sys.exc_info()[0]
+    print '4', sys.exc_info()[1]
+    print '5', traceback.tb_lineno(sys.exc_info()[2])
+    ex_type, ex, tb = sys.exc_info()
+    print '6', traceback.print_tb(tb)
